Not perform request if schema output exists

+++ dts-downloader/src/lib.rs
@@ -50,13 +50,15 @@ pub fn run_transform(
         out_path_rs: OutPathRs(rs_file),
     }: Opt,
 ) -> Result<()> {
-    // setup .d.ts file
-    let repo = "octokit/webhooks";
-    let url =
-        format!("https://raw.githubusercontent.com/{repo}/{branch}/payload-types/schema.d.ts");
+    if !dts_file.try_exists()? {
+        // setup .d.ts file
+        let repo = "octokit/webhooks";
+        let url =
+            format!("https://raw.githubusercontent.com/{repo}/{branch}/payload-types/schema.d.ts");
 
-    let body = reqwest::blocking::get(url)?.text()?;
-    std::fs::write(&dts_file, body)?;
+        let body = reqwest::blocking::get(url)?.text()?;
+        std::fs::write(&dts_file, body)?;
+    }
 
     let rs = dts2rs(&dts_file);
